requirements.txt
fastapi==0.121.2
click==8.1.7
tortoise-orm==0.25.1
asyncpg==0.30.0
bcrypt==4.3.0
passlib==1.7.4
python-jose==3.5.0
python-dotenv==1.1.1
uvicorn==0.35.0
uvloop==0.21.0
pytest==9.0.1
httpx==0.28.1
pydantic[email]
